Update to later code from my repository: o many bug fixes o add new periodic calibration api o break up 5416 periodic calibration code in preparation for 928x o move get noise floor to rf backends o 5416-specific ani (still disabled) o modularize 5210 eeprom format a la other eeprom formats o start cleaning up regdomain code o prepare for proper 1/2 and 1/4 width channel support o bring back 900MHz card support o clean up 5212 rf version handling o add 1/2 and 1/4 width channel support for 5212 parts o split 5212 rfgain handling out o improve ani debugging o add AH_USE_INIPDGAIN compile option o purge a bunch of dead 5212 state o add 1/2 and 1/4 rate modes o remove HAL_CAP_CHAN_HALFRATE and HAL_CAP_CHAN_QUARTERRATE; the   same info can now be deduced from the set of supported modes

+++ ah_eeprom_v14.c
@@ -14,7 +14,7 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
- * $Id: ah_eeprom_v14.c,v 1.3 2008/11/10 04:08:00 sam Exp $
+ * $Id: ah_eeprom_v14.c,v 1.4 2008/11/10 19:04:26 sam Exp $
  */
 #include "opt_ah.h"
 
@@ -37,10 +37,10 @@ v14EepromGet(struct ath_hal *ah, int param, void *val)
 
 	switch (param) {
         case AR_EEP_NFTHRESH_5:
-		*(int8_t *)val = pModal[0].noiseFloorThreshCh[0];
+		*(int16_t *)val = pModal[0].noiseFloorThreshCh[0];
 		return HAL_OK;
         case AR_EEP_NFTHRESH_2:
-		*(int8_t *)val = pModal[1].noiseFloorThreshCh[0];
+		*(int16_t *)val = pModal[1].noiseFloorThreshCh[0];
 		return HAL_OK;
         case AR_EEP_MACADDR:		/* Get MAC Address */
 		sum = 0;
